testing again
When GHA are triggered using PR they run the code that already exists in main.
Need to test the action that are now  in the dev branch.  Added code to the
action to address network interput issues when running curl.  Modfy the trigger on the action to
test the retry on that has been implemented on the curl call that
downloads the wgrib dependencies.